backend/requirements.txt
django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.4,<5.6

channels>=4.0,<4.1
channels-redis>=4.2,<4.3
daphne>=4.1,<4.2

celery>=5.4,<5.5
redis>=5.0,<5.1
django-redis>=5.4,<5.5

psycopg[binary]>=3.1,<3.2

python-dotenv>=1.0,<2.0